修正物品拾取99以适配客户端 修改数据库操作模型

+++ 数据库/auth_query.c
@@ -22,7 +22,7 @@
 /* 数据库连接 */
 extern psocn_dbconn_t conn;
 
-int db_upload_temp_data(void* data, uint32_t size) {
+int db_upload_temp_data(void* data, size_t size) {
 
     memset(myquery, 0, sizeof(myquery));
